Move useCurrentHexathon file
